requirements.txt
django==5.0.6
dash==2.17.0
dash-bootstrap-components==1.6.0
# psutil conflict: reflex needs psutil>=7.0.0,<8.0, safety needs ~=6.1.0. If you need to run safety, consider a separate environment.
uvicorn==0.29.0
daphne==4.1.2
python-dotenv==1.0.1
djangorestframework==3.15.1
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.4
windpowerlib==0.2.2
jax
pymodbus==3.9.2
opcua==0.98.13
paho-mqtt==2.1.0
tensorflow
matplotlib==3.9.0
plotly==5.22.0
numpy==1.26.4
pandas==2.2.2
channels==4.0.0
whitenoise==6.7.0
